backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "MidnightTasks backend"
authors = [
    "SmokingA1 <denisshokolad1221@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.42"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
alembic = "^1.16.4"
asyncpg = "^0.30.0"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
setuptools = "^80.9.0"
redis = "^6.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"